scripts/requirements.txt
httpx
pydantic
youtube-transcript-api
google-genai
requests
beautifulsoup4
